Add base MESolver class
